Generate HTML on the fly
Instead of storing all run data and generating the HTML in Finalize,
this implementation generates the necessary HTML snippets in ReportRuns.
On Finalize the strings are concatenated and replaced.

Skip output of empty line/bar charts: Move chart specific strings to
ChartIt: This refactoring makes it simple to change the order of the
charts. The strings that belong together are grouped explicitly except
of implicitly via the array index.

Reorder charts to Real, CPU, items, Bytes: This is less surprising than
the back and forth between [s] and [s^-1]

Add an initial HTMLReporter test

Report scaling and debug information in the HTML footer

Use recursive hashmap based HTML templating

* Modified tooltips to combine all values at a given x
* Print stddev in tooltip if available
* get rid of some hardcoded strings by relying only on BENCHMARK_NAME
  and UNIT
* Print only 3 decimals in the tooltip
* Enable scrollbar/navigator for bar charts
* Make charts 700px large
* Go over every HTML snippet only once, replacing all ${TAGS} as they
  occur. Lookup of the replacement is done via a unordered_map argument
* To avoid multiple passes of tag searching, recursively build the chart
  snippets and insert the result via a tag replacement
* Use ReplaceHTMLSpecialChars to generate a correct display name

Signed-off-by: Matthias Kretz <kretz@kde.org>