Generate title from arg0, /proc/cpuinfo and macros
Generate the Context::title from argc[0], the CPU vendor, family, and
model information from /proc/sysinfo, and the predefined compiler macros
expanded at compile time of the user application. Thus, the title string
reports the executable name, host CPU id, and compiler used to compile
the benchmark application.

Print the Context::title as <h1> on the HTML output on top (in addition
to using it in <title>). Also move the rest of the context information
to the top as well.

Signed-off-by: Matthias Kretz <kretz@kde.org>